u to size_t to avoid truncation
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@1106 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/phycas_string.cpp
@@ -153,7 +153,7 @@ string GetSingleQuoted(const string & f)
 unsigned replace_all_substr(std::string &target, const std::string & searchStr, const std::string & replaceStr)
 	{
 	unsigned nReplacements = 0;
-	unsigned i = (unsigned)target.find(searchStr, 0);
+	size_t i = (unsigned)target.find(searchStr, 0);
 	while (i != string::npos)
 		{
 		++nReplacements;
